src/requirements.txt
APScheduler==3.7.0
asgiref==3.3.0
beautifulsoup4==4.9.1
certifi==2020.12.5
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-import-export==2.3.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.1.0
Pillow==7.2.0
psycopg2==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
rpa==1.27.1
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
tagui==1.27.1
tzlocal==2.1
urllib3==1.25.11
xlrd==2.0.1
xlwt==1.3.0
